HeppyCore: allow splitting a file in multiple chunks
This feature can be enabled by setting fineSplitFactor = N > 1 on
a component. In that case, the scheduler will make N jobs for each
file, and each job will process 1/N of the events in the file.

To avoid code duplication, now heppy_batch takes the split function
from heppy.

Also, the print out of the time per event has been updated in order
to work correctly with this feature, or in general when not processing
events starting from the first one in the file.